requirements.txt
python-docx
pillow